requirements.txt
absl-py                      2.1.0
addict                       2.4.0
aliyun-python-sdk-core       2.14.0
aliyun-python-sdk-kms        2.16.2
antlr4-python3-runtime       4.9.3
astunparse                   1.6.3
cachetools                   5.3.3
certifi                      2023.11.17
cffi                         1.16.0
charset-normalizer           3.3.2
click                        8.1.7
colorama                     0.4.6
contourpy                    1.1.1
crcmod                       1.7
cryptography                 41.0.7
cycler                       0.12.1
Cython                       3.0.10
einops                       0.8.0
filelock                     3.13.1
flatbuffers                  24.3.25
fonttools                    4.47.0
fsspec                       2023.12.2
ftfy                         6.1.3
gast                         0.4.0
google-auth                  2.31.0
google-auth-oauthlib         1.0.0
google-pasta                 0.2.0
grpcio                       1.64.1
h5py                         3.11.0
huggingface-hub              0.20.2
idna                         3.6
imageio                      2.33.1
importlib-metadata           7.0.1
importlib-resources          6.1.1
Jinja2                       3.1.2
jmespath                     0.10.0
joblib                       1.4.2
jsonpatch                    1.33
jsonpointer                  3.0.0
keras                        2.13.1
kiwisolver                   1.4.5
kornia                       0.7.3
kornia_rs                    0.1.3
lazy_loader                  0.3
libclang                     18.1.1
Markdown                     3.5.1
markdown-it-py               3.0.0
MarkupSafe                   2.1.3
matplotlib                   3.7.4
mdurl                        0.1.2
mmcv                         2.1.0
mmengine                     0.10.3
model-index                  0.1.11
mpmath                       1.3.0
natsort                      8.4.0
networkx                     3.1
numpy                        1.24.3
nvidia-cublas-cu12           12.1.3.1
nvidia-cuda-cupti-cu12       12.1.105
nvidia-cuda-nvrtc-cu12       12.1.105
nvidia-cuda-runtime-cu12     12.1.105
nvidia-cudnn-cu12            8.9.2.26
nvidia-cufft-cu12            11.0.2.54
nvidia-curand-cu12           10.3.2.106
nvidia-cusolver-cu12         11.4.5.107
nvidia-cusparse-cu12         12.1.0.106
nvidia-nccl-cu12             2.18.1
nvidia-nvjitlink-cu12        12.3.101
nvidia-nvtx-cu12             12.1.105
oauthlib                     3.2.2
omegaconf                    2.3.0
opencv-python                4.9.0.80
opencv-python-headless       4.9.0.80
opendatalab                  0.0.10
openmim                      0.3.9
openxlab                     0.0.33
opt-einsum                   3.3.0
ordered-set                  4.1.0
oss2                         2.17.0
packaging                    23.2
pandas                       2.0.3
pillow                       10.2.0
pip                          23.3.2
prettytable                  3.9.0
protobuf                     4.25.3
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pycocotools                  2.0.7
pycparser                    2.21
pycryptodome                 3.19.1
Pygments                     2.17.2
pyparsing                    3.1.1
python-dateutil              2.8.2
pytorch-msssim               1.0.0
pytz                         2023.3.post1
PyWavelets                   1.4.1
PyYAML                       6.0.1
regex                        2023.12.25
requests                     2.28.2
requests-oauthlib            2.0.0
rich                         13.4.2
rsa                          4.9
safetensors                  0.4.1
scikit-image                 0.21.0
scikit-learn                 1.3.2
scipy                        1.10.1
seaborn                      0.13.2
setuptools                   60.2.0
shapely                      2.0.3
six                          1.16.0
sympy                        1.12
tabulate                     0.9.0
tensorboard                  2.13.0
tensorboard-data-server      0.7.2
termcolor                    2.4.0
terminaltables               3.1.10
threadpoolctl                3.5.0
tifffile                     2023.7.10
timm                         0.9.12
torch                        2.1.2+cu118
torchaudio                   2.1.2
torchvision                  0.16.2
tornado                      6.4.1
tqdm                         4.65.2
triton                       2.1.0
typing_extensions            4.5.0
tzdata                       2023.4
urllib3                      1.26.18
visdom                       0.2.4
wcwidth                      0.2.13
websocket-client             1.8.0
Werkzeug                     3.0.3
wheel                        0.42.0
wrapt                        1.16.0
yapf                         0.32.0
zipp                         3.17.0
